--- app/src/main/java/hpsaturn/pollutionreporter/report/domain/entities/Exceptions.kt
@@ -1,3 +1,0 @@
-package hpsaturn.pollutionreporter.report.domain.entities
-
-class TracksInfoNotFoundException : Throwable()
