+++ app/src/main/java/hpsaturn/pollutionreporter/reports/open/domain/entities/Exceptions.kt
@@ -1,0 +1,3 @@
+package hpsaturn.pollutionreporter.reports.open.domain.entities
+
+class TracksInfoNotFoundException : Throwable()
