+++ app/src/main/res/layout/fragment_dashboard.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".dashboard.DashboardFragment">
+    tools:context=".dashboard.presentation.DashboardFragment">
 
 
     <ImageView
